[nrf noup] ci: Add additional sauce tag Git linting
Add an extra step in the compliance workflow to check sauce tags as well
as formatting of commit messages based on sauce tag.

In order to not interfere with CI in other repos (namely sdk-nrf) this
is run as an independent gitlint invocation and not as part of
check_compliance.py. This invocation comes with its own .gitlint file
due to limitations in gitlint itself.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>